+++ forui/CHANGELOG.md
@@ -334,8 +334,10 @@ Unfortunately, `FCalendar` cannot be easily migrated using data driven fixes due
 
 ### `FTimeField`
 * Add `FTimeFieldControl`.
+* Add `FTimeField.picker(groupId: ...)`.
 
 * **Breaking** Rename `FTimeField.shift` to `FTimeField.overflow`.
+* **Breaking** Rename `FTimeField.picker(inputAnchor: ...)` to `FTimeField.picker(fieldAnchor: ...)`.
 * **Breaking** Remove `FTimeField.controller`. Use `FTimeField(control: .managed(...))` instead.
 * **Breaking** Remove `FTimeField.onChange`. Use `FTimeField(control: .managed(...))` instead.
 * **Breaking** Remove `FTimeField.initialTime`. Use `FTimeField(control: .managed(...))` instead.